fix : git 충돌 오류 수정

+++ molaeng_back/src/main/java/idle/molaeng_back/user/model/GugunRepository.java
@@ -3,6 +3,6 @@ package idle.molaeng_back.user.model;
 import io.swagger.models.auth.In;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface GugunRepository extends JpaRepository<Gugun, Integer> {
+public interface GugunRepository extends JpaRepository<Gugun, Long> {
     Gugun findByGugunName(String gugunName);
 }
